feat: refactored export utils to support new and old approaches (M2-7496)
