+++ src/shared/utils/exportData/exportDataSucceed.ts
@@ -5,15 +5,50 @@ import {
   reportHeader,
 } from 'shared/consts';
 import { useDecryptedActivityData } from 'modules/Dashboard/hooks';
-import { ExportDataResult } from 'shared/types';
+import {
+  AppletExportData,
+  ExportDataResult,
+  ExtendedExportAnswerWithoutEncryption,
+  DecryptedActivityData,
+} from 'shared/types';
 
 import { exportTemplate } from '../exportTemplate';
 import { exportCsvZip } from './exportCsvZip';
 import { exportMediaZip } from './exportMediaZip';
 import { getReportZipName, ZipFile } from './getReportName';
-import { ExportDataFilters, prepareData } from './prepareData';
+import { ExportDataFilters, prepareEncryptedData, prepareDecryptedData } from './prepareData';
 
-export const exportDataSucceed =
+const exportProcessedData = async ({
+  reportData,
+  activityJourneyData,
+  mediaData,
+  drawingItemsData,
+  stabilityTrackerItemsData,
+  abTrailsItemsData,
+  flankerItemsData,
+  suffix,
+}: AppletExportData & { suffix: string }) => {
+  await exportTemplate({
+    data: reportData,
+    fileName: GENERAL_REPORT_NAME + suffix,
+    defaultData: reportData.length > 0 ? null : reportHeader,
+  });
+  await exportTemplate({
+    data: activityJourneyData,
+    fileName: JOURNEY_REPORT_NAME + suffix,
+    defaultData: activityJourneyData.length > 0 ? null : activityJourneyHeader,
+  });
+
+  await Promise.allSettled([
+    exportCsvZip(drawingItemsData, getReportZipName(ZipFile.Drawing, suffix)),
+    exportCsvZip(stabilityTrackerItemsData, getReportZipName(ZipFile.StabilityTracker, suffix)),
+    exportCsvZip(abTrailsItemsData, getReportZipName(ZipFile.ABTrails, suffix)),
+    exportCsvZip(flankerItemsData, getReportZipName(ZipFile.Flanker, suffix)),
+    exportMediaZip(mediaData, getReportZipName(ZipFile.Media, suffix)),
+  ]);
+};
+
+export const exportEncryptedDataSucceed =
   ({
     getDecryptedAnswers,
     suffix,
@@ -26,32 +61,15 @@ export const exportDataSucceed =
   async (result: ExportDataResult) => {
     if (!result) return;
 
-    const {
-      reportData,
-      activityJourneyData,
-      mediaData,
-      drawingItemsData,
-      stabilityTrackerItemsData,
-      abTrailsItemsData,
-      flankerItemsData,
-    } = await prepareData(result, getDecryptedAnswers, filters);
+    const exportData = await prepareEncryptedData(result, getDecryptedAnswers, filters);
+    await exportProcessedData({ ...exportData, suffix });
+  };
 
-    await exportTemplate({
-      data: reportData,
-      fileName: GENERAL_REPORT_NAME + suffix,
-      defaultData: reportData.length > 0 ? null : reportHeader,
-    });
-    await exportTemplate({
-      data: activityJourneyData,
-      fileName: JOURNEY_REPORT_NAME + suffix,
-      defaultData: activityJourneyData.length > 0 ? null : activityJourneyHeader,
-    });
+export const exportDecryptedDataSucceed =
+  ({ suffix, filters }: { suffix: string; filters?: ExportDataFilters }) =>
+  async (parsedAnswers: DecryptedActivityData<ExtendedExportAnswerWithoutEncryption>[]) => {
+    if (!parsedAnswers) return;
 
-    await Promise.allSettled([
-      exportCsvZip(drawingItemsData, getReportZipName(ZipFile.Drawing, suffix)),
-      exportCsvZip(stabilityTrackerItemsData, getReportZipName(ZipFile.StabilityTracker, suffix)),
-      exportCsvZip(abTrailsItemsData, getReportZipName(ZipFile.ABTrails, suffix)),
-      exportCsvZip(flankerItemsData, getReportZipName(ZipFile.Flanker, suffix)),
-      exportMediaZip(mediaData, getReportZipName(ZipFile.Media, suffix)),
-    ]);
+    const exportData = await prepareDecryptedData(parsedAnswers, filters);
+    await exportProcessedData({ ...exportData, suffix });
   };
